Validação parte 1 - controle de fluxos, aula 299

+++ app/Http/Controllers/MarcaController.php
@@ -56,6 +56,9 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Recurso pesquisado não existe'];
+        }
         return $marca;
     }
 
@@ -81,6 +84,10 @@ class MarcaController extends Controller
     {
         //$marca ->update(($request->all()));
         $marca = $this->marca->find($id);
+
+        if($marca === null){
+            return ['erro' => 'Impossivel realizar a atualização. O recurso solicitado não existe'];
+        }
         $marca->update($request->all());
         return $marca;
     }
@@ -94,6 +101,10 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+
+        if($marca === null){
+            return ['erro' => 'Impossivel realizar a exclusão. O recurso solicitado não existe'];
+        }
         $marca->delete();
         return ['msg' => 'A marca foi removida com sucesso!'];
     }
